validacion rutas jornadas

+++ controllers/JornadaController.js
@@ -1,4 +1,6 @@
 import Jornada from "../models/Jornada.js";
+import JornadaCicloEscolar from "../models/JornadaCicloEscolar.js";
+import { decodeJWT } from "../utils/codificar.js";
 
 export const getAllJornadas = async (req, res) => {
     try {
@@ -37,6 +39,15 @@ export const getJornadaByNivelEducacion = async (req, res) => {
 
 export const createJornada = async (req, res) => {
     try {
+        const token = req.cookies?.token;
+        const userData = decodeJWT(token);
+
+        if (!userData) {
+            return res.status(403).json({ status: 'ERROR', message: "Token inválido o no proporcionado" });
+        }
+
+        req.body.id_usuario_creo = userData.id
+
         await Jornada.create(req.body);
         res.status(200).json({ status: 'OK', message: 'Jornada creada correctamente!' });
     } catch (error) {
@@ -46,6 +57,15 @@ export const createJornada = async (req, res) => {
 
 export const updateJornada = async (req, res) => {
     try {
+        const token = req.cookies?.token;
+        const userData = decodeJWT(token);
+
+        if (!userData) {
+            return res.status(403).json({ status: 'ERROR', message: "Token inválido o no proporcionado" });
+        }
+
+        req.body.id_usuario_modifico = userData.id
+
         await Jornada.update(req.body, {
             where: {
                 id_jornada: req.params.id
@@ -63,6 +83,12 @@ export const deleteJornada = async (req, res) => {
 
         if (!jornada) {
             return res.status(400).json({ status: 'ERROR', message: 'Jornada no encontrada' });
+        }
+
+        const jornadaCiclo = await JornadaCicloEscolar.findOne({ where: { id_jornada: req.params.id }})
+
+        if (jornadaCiclo) {
+            return res.status(400).json({ status: 'ERROR', message: 'Jornada asociada a un ciclo escolar' });
         }
 
         await Jornada.destroy({
